Minor name fix & removed log file from track list.
The terminationbyte name was not right, due to it is meaning the size of newline.
It is necessary for calculation of file size. I do not want to read the file length
in every times when I use it. This read is enought at begin of file usage, after
that it increments internally with the size of output message & with this size of newline.